requirements.txt
flask==2.3.2
flask_sqlalchemy==3.1.1
python_dotenv==1.0.0
flask_limiter==3.5.0